Fix for #48911 + rebase to https://github.com/JuliaLang/julia/commit/219dc107dabe24a3a99b2e9a5b33b777437cc27b

Summary:

1. Fix the application of the elementary transformation (correct complex conjugation) to the solution as per LAPACK zlarzb https://netlib.org/lapack/explore-html/d5/ddd/zlarzb_8f_source.html
2. Fix a complex conjugate bug in the original solution for https://github.com/JuliaLang/julia/issues/48911, align with LAPACK zlarzb, https://netlib.org/lapack/explore-html/d5/ddd/zlarzb_8f_source.html
3. ~~Fix output permutation when A.P is not the identity matrix.~~ (This aspect was removed during integration as the permutation is taken care by `ldiv!`. Previously the code was tested in isolation.)
4. ~~Align the final permutation with LAPACK's ZGELSY at https://netlib.org/lapack/explore-html/d8/d6e/zgelsy_8f_source.html~~ (This aspect was removed during integration)
5. ~~Improve permutation performance~~ (This aspect was removed during integration)