@aravindh-krishnamoorthy 2023-05-02
@N5N3 The tests are included. Thanks! The tests helped me fix a bug. The new test is called "issue #48911" and is located at the end. Here's the report (run in the LinearAlgebra/test directory):

```julia
julia> using Revise
julia> using LinearAlgebra
julia> Revise.track(LinearAlgebra)
julia> include("qr.jl")
Test Summary:   | Pass  Total   Time
eltya = Float32 |  740    740  12.1s
Test Summary:   | Pass  Total  Time
eltya = Float64 |  740    740  4.7s
Test Summary:      | Pass  Total  Time
eltya = ComplexF32 |  740    740  5.5s
Test Summary:      | Pass  Total  Time
eltya = ComplexF64 |  740    740  4.1s
Test Summary:    | Pass  Total   Time
eltya = BigFloat |  740    740  10.8s
Test Summary: | Pass  Total  Time
eltya = Int64 |  500    500  1.8s
Test Summary:    | Pass  Total  Time
transpose errors |    3      3  0.2s
Test Summary: | Pass  Total  Time
Issue 7304    |    1      1  0.0s
Test Summary:        | Pass  Total  Time
qr on AbstractVector |    8      8  0.4s
Test Summary: | Pass  Total  Time
QR on Ints    |    1      1  0.2s
Test Summary: | Pass  Total  Time
Issue 16520   |    1      1  0.1s
Test Summary: | Pass  Total  Time
Issue 22810   |    2      2  0.1s
Test Summary: | Pass  Total  Time
Issue 24107   |    1      1  0.2s
Test Summary:                               | Pass  Total  Time
Issue 24589. Promotion of rational matrices |    1      1  0.1s
Test Summary:                                               | Pass  Total  Time
Issue Test Factorization fallbacks for rectangular problems |    8      8  0.1s
Test Summary:                         | Pass  Total  Time
Issue reflector of zero-length vector |    3      3  0.2s
Test Summary:                          | Pass  Total  Time
det(Q::Union{QRCompactWYQ, QRPackedQ}) |  376    376  0.5s
Test Summary:    | Pass  Total  Time
inv(::AbstractQ) |    4      4  0.0s
Test Summary:         | Pass  Total  Time
QR factorization of Q |    8      8  0.1s
Test Summary:                     | Pass  Total  Time
Generation of orthogonal matrices |    2      2  0.2s
Test Summary:                           | Pass  Total  Time
Multiplication of Q by special matrices |   16     16  0.6s
Test Summary: | Pass  Total  Time
copyto! for Q |   12     12  1.9s
Test Summary: | Pass  Total  Time
adjoint of QR |   48     48  0.0s
Test Summary: | Pass  Total  Time
issue #38974  |    2      2  0.1s
Test Summary:           | Pass  Total  Time
convert between eltypes |    6      6  1.0s
Test Summary:                       | Pass  Total  Time
optimized getindex for an AbstractQ |   79     79  0.7s
Test Summary: | Pass  Total  Time
issue #48911  |    2      2  0.8s
Main.TestQR
```