@dkarrasch 2023-05-04
I'd suggest to expicitly including the original example as a test case, and test that the solution is close to `[1 - 0.5im, 0.5 + 1im]`. In fact, I believe this test might be sufficient to add, because it reveals all those fine mistakes we had earlier.