@dkarrasch 2023-05-04
Also, could you please update the initial summary. There are listed many things, but in the end it's "just" a few `'` here and there, and nothing related to permutations, for instance.